Increase DelayedJob timeout & fix serialize bug
If a job took longer than 4 hours to run a Timeout exception would be
raised. An attempt would be made to serialize the exception as a
TransporterError, causing the worker to exit and the failed job to
remain in the running state until the worker was restarted.

This commit increas the timeout to 48 hours